feature-api-call - add env file and utils, move canned data into util
