+++ .env
@@ -1,0 +1,2 @@
+# API endpoint
+VITE_FED_API='http://localhost:3000/api'
